coolsite/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-debug-toolbar==3.5.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
